pyproject.toml
[project]
name = "kinetixverse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chumpy",
    "faiss-cpu>=1.13.2",
    "huggingface-hub>=1.5.0",
    "lightglue",
    "mediapipe==0.10.9",
    "moviepy>=2.2.1",
    "numpy>=2.4.2",
    "ollama>=0.6.1",
    "open3d>=0.19.0",
    "opencv-python>=4.13.0.92",
    "rerun-sdk>=0.30.1",
    "torch>=2.10.0",
    "trackers>=2.2.0",
    "transformers>=5.2.0",
    "trimesh>=4.11.3",
    "ultralytics>=8.4.19",
]

[tool.uv.sources]
chumpy = { git = "https://github.com/hassony2/chumpy.git" }
lightglue = { git = "https://github.com/cvg/LightGlue.git" }

[tool.uv.workspace]
members = [
    "mvp/manopth",
]
